Sound: fix legacy Wav load/save

+++ src/core/vpinball.cpp
@@ -2173,8 +2173,8 @@ int CALLBACK MyCompProcIntValues(LPARAM lSortParam1, LPARAM lSortParam2, LPARAM 
 
    std::string buf1(256, '\0');
    std::string buf2(256, '\0');
-   ListView_GetItemText(lpsd->hwndList, nItem1, lpsd->subItemIndex, &buf1[0], buf1.size());
-   ListView_GetItemText(lpsd->hwndList, nItem2, lpsd->subItemIndex, &buf2[0], buf2.size());
+   ListView_GetItemText(lpsd->hwndList, nItem1, lpsd->subItemIndex, &buf1[0], (int)buf1.size());
+   ListView_GetItemText(lpsd->hwndList, nItem2, lpsd->subItemIndex, &buf2[0], (int)buf2.size());
    int value1 = 0, value2 = 0;
    try_parse_int(buf1, value1);
    try_parse_int(buf2, value2);
